backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
google-genai>=1.0.0
newsapi-python==0.2.7
pydantic==2.5.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
